Allow trace filename to be overridden with server hostname.
This change introduces a boolean option `override_hostname` in
`ProfileOptions.advanced_configuration`. If this option is set to true,
the profiler server will use `tsl::port::Hostname()` as the filename
for saving the XPlane trace, instead of the host identifier provided
in the request. This allows for more predictable trace filenames in
environments like GKE where host IPs may not be stable or meaningful.

PiperOrigin-RevId: 834235467